fixes layout
