+++ src/app/tienda/layout.jsx
@@ -5,16 +5,9 @@ import { config } from "@fortawesome/fontawesome-svg-core";
 
 config.autoAddCss = false;
 
-export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-  brandName: 'Barb & Heimer',
-}
-
-export default function RootLayout({ children }) {
+export default function NavBarLayout({ children }) {
   return (
     <div>
-      <div className="brand-name">{metadata.brandName}</div>
       <div className="navigation-bar">
         <p className="texto-navigation-bar">Todos</p>
         <p className="texto-navigation-bar">Ojos</p>
